requirements.txt
opencv_python==4.5.3.56
streamlit==0.88.0
numpy==1.21.2
deepface==0.0.68
Pillow==8.3.2
